+++ .github/workflows/vitest.yml
@@ -7,6 +7,7 @@ on:
     paths:
       - "src/**"
       - "musea/**"
+      - "config/**"
       - "types/**"
       - "package.json"
       - "pnpm-workspace.yaml"
