--- musea/preview.css
@@ -1,51 +1,0 @@
-body {
-  background: var(--paper);
-  padding: clamp(16px, 2vw, 28px);
-}
-
-.musea-variant {
-  background: var(--paper);
-  color: var(--ink);
-  min-height: calc(100vh - clamp(32px, 4vw, 56px));
-}
-
-.musea-variant,
-.musea-variant * {
-  margin: revert-layer;
-  padding: revert-layer;
-}
-
-.musea-preview-compact,
-.musea-preview-wide,
-.musea-preview-page {
-  background: var(--paper);
-  color: var(--ink);
-  width: 100%;
-}
-
-.musea-preview-page {
-  min-height: 680px;
-}
-
-.musea-dark-surface {
-  --paper: #1c1c19;
-  --paper-2: #262622;
-  --rule: #e8e4dc;
-  --rule-soft: rgba(232, 228, 220, 0.15);
-  --rule-softer: rgba(232, 228, 220, 0.07);
-  --ink: #e8e4dc;
-  --ink-2: #b4b0a8;
-  --ink-3: #7a7a72;
-  --accent: #2ed9a8;
-  --accent-ink: #0d2820;
-  --color-paper: var(--paper);
-  --color-paper-2: var(--paper-2);
-  --color-rule: var(--rule);
-  --color-rule-soft: var(--rule-soft);
-  --color-rule-softer: var(--rule-softer);
-  --color-ink: var(--ink);
-  --color-ink-2: var(--ink-2);
-  --color-ink-3: var(--ink-3);
-  --color-accent: var(--accent);
-  --color-accent-ink: var(--accent-ink);
-}
