+++ musea/theme.ts
@@ -1,0 +1,22 @@
+export const museaDarkSurfaceStyle = {
+  "--paper": "#1c1c19",
+  "--paper-2": "#262622",
+  "--rule": "#e8e4dc",
+  "--rule-soft": "rgba(232, 228, 220, 0.15)",
+  "--rule-softer": "rgba(232, 228, 220, 0.07)",
+  "--ink": "#e8e4dc",
+  "--ink-2": "#b4b0a8",
+  "--ink-3": "#7a7a72",
+  "--accent": "#2ed9a8",
+  "--accent-ink": "#0d2820",
+  "--color-paper": "var(--paper)",
+  "--color-paper-2": "var(--paper-2)",
+  "--color-rule": "var(--rule)",
+  "--color-rule-soft": "var(--rule-soft)",
+  "--color-rule-softer": "var(--rule-softer)",
+  "--color-ink": "var(--ink)",
+  "--color-ink-2": "var(--ink-2)",
+  "--color-ink-3": "var(--ink-3)",
+  "--color-accent": "var(--accent)",
+  "--color-accent-ink": "var(--accent-ink)",
+};
