+++ src/components/AppHeader.art.vue
@@ -1,4 +1,6 @@
 <script setup lang="ts">
+import { museaDarkSurfaceStyle } from "../../musea/theme";
+
 defineArt("./AppHeader.vue", {
   title: "AppHeader",
   category: "Layout",
@@ -8,12 +10,12 @@ defineArt("./AppHeader.vue", {
 
 <art>
   <variant name="Default" default>
-    <div class="musea-preview-wide min-h-[96px] bg-paper text-ink">
+    <div class="min-h-[96px] bg-paper text-ink">
       <Self />
     </div>
   </variant>
   <variant name="Dark surface">
-    <div class="musea-dark-surface musea-preview-wide min-h-[96px] bg-paper text-ink">
+    <div :style="museaDarkSurfaceStyle" class="min-h-[96px] bg-paper text-ink">
       <Self />
     </div>
   </variant>
