+++ src/components/AppLogoMark.art.vue
@@ -1,4 +1,6 @@
 <script setup lang="ts">
+import { museaDarkSurfaceStyle } from "../../musea/theme";
+
 defineArt("./AppLogoMark.vue", {
   title: "AppLogoMark",
   category: "Components",
@@ -14,7 +16,8 @@ defineArt("./AppLogoMark.vue", {
   </variant>
   <variant name="Dark surface">
     <div
-      class="musea-dark-surface min-h-[160px] flex items-center justify-center bg-paper text-accent"
+      :style="museaDarkSurfaceStyle"
+      class="min-h-[160px] flex items-center justify-center bg-paper text-accent"
     >
       <Self class="w-12 h-auto" />
     </div>
