+++ tsconfig.json
@@ -18,6 +18,7 @@
     "src/**/*.vue",
     "musea/**/*.ts",
     "types/**/*.ts",
+    "config/**/*.ts",
     "env.d.ts",
     "vize.config.ts",
     "vite.config.ts"
