+++ tsconfig.vize.json
@@ -8,7 +8,8 @@
     "./src/**/*.vue",
     "./src/data/**/*.ts",
     "./musea/**/*.ts",
-    "./types/**/*.ts"
+    "./types/**/*.ts",
+    "./config/**/*.ts"
   ],
   "exclude": [
     "./src/**/*.test.ts",
